+++ test/test_models_scrabble.py
@@ -4,8 +4,6 @@ from game.models import (
     Tile,
     Board,
     Cell,
-
-
 )
 from unittest.mock import patch
 
@@ -14,6 +12,7 @@ class TestTiles(unittest.TestCase):
         tile = Tile('A', 1)
         self.assertEqual(tile.letter, 'A')
         self.assertEqual(tile.value, 1)
+
 class TestCalculateWordValue(unittest.TestCase):
     def test_word_nomult(self):
         word = [
@@ -41,13 +40,14 @@ class TestCalculateWordValue(unittest.TestCase):
                     ]           
         result = Board.calculatewordvalue(word)
         self.assertEqual(result, 5)
+
 class TestBagTiles(unittest.TestCase):
     @patch('random.shuffle')
     def test_bag_tiles(self, patch_shuffle):
         bag = BagTiles()
         self.assertEqual(
             len(bag.finaltiles),
-            100,
+            103,
         )
         self.assertEqual(
             patch_shuffle.call_count,
@@ -62,23 +62,26 @@ class TestBagTiles(unittest.TestCase):
         tiles = bag.take(2)
         self.assertEqual(
             len(bag.finaltiles),
-            98,
+            101,
         )
         self.assertEqual(
             len(tiles),
             2,
         )
+
     def test_put(self):
         bag = BagTiles()
         put_tiles = [('Z', 1), ('Y', 1)]
         bag.put(put_tiles)
         self.assertEqual(
             len(bag.finaltiles),
-            102,
+            105,
         )
+
     def test_board(self):
         board = Board()
         self.assertEqual(len(board.grid), 15 )
+
     def test_cell(self):
         celda = Cell(2, 'letter', ('',0))
         letra = ('A', 1)
@@ -87,13 +90,16 @@ class TestBagTiles(unittest.TestCase):
         celda.add_letter(letra)
         self.assertEqual(celda.letter.letter, 'A')
         self.assertEqual(celda.letter.value, 1)
+
     def test_calculatevalue0(self):
         celda2 = Cell(1,'',('',0))
         self.assertEqual(celda2.calculate_value(), 0)
+
     def test_calculatevaluemultilpierletter(self):
         celda3 = Cell(2, 'letter',('',0))
         celda3.add_letter(('P', 3))
         self.assertEqual(celda3.calculate_value(), 6)
+
     def test_calculatevaluemultilpierword(self):
         celda4 = Cell(2, 'Word',('',0))
         celda4.add_letter(('P', 3))