+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 
 ### Added
 - New test for the class Player, checking for different possibilities
+- New main.py method that asks the player for inputs (2023/10/17)
+-
+
+### Fixed
+- Errors caused by deleting double-letter words (2023/10/17)
+
 
 ## [V 1.1.5] 2023/09/27: